Move LibrariesHandler handling code and watchers to LibraryManager
