+++ src/Files.App/Filesystem/LibraryManager.cs
@@ -1,20 +1,39 @@
 using Files.Shared;
-using Files.App.Helpers;
+using Files.Shared.Enums;
 using Files.App.Extensions;
+using Files.App.ViewModels.Dialogs;
+using Files.App.Dialogs;
+using Files.App.Shell;
+using Microsoft.UI.Xaml.Controls;
 using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Windows.System;
+using Visibility = Microsoft.UI.Xaml.Visibility;
+using Vanara.PInvoke;
+using Vanara.Windows.Shell;
+using Files.Shared.Extensions;
+using static Vanara.PInvoke.CM_PARTIAL_RESOURCE_DESCRIPTOR;
+using Windows.Foundation.Collections;
+using System.Text.Json;
+using Files.App.Helpers;
 
 namespace Files.App.Filesystem
 {
     public class LibraryManager : IDisposable
     {
-        public EventHandler<NotifyCollectionChangedEventArgs> DataChanged;
+        public EventHandler<NotifyCollectionChangedEventArgs>? DataChanged;
 
+        private FileSystemWatcher librariesWatcher;
         private readonly List<LibraryLocationItem> libraries = new();
+        private static readonly Lazy<LibraryManager> lazy = new(() => new LibraryManager());
+
+        public static LibraryManager Default
+            => lazy.Value;
+
         public IReadOnlyList<LibraryLocationItem> Libraries
         {
             get
@@ -26,13 +45,91 @@ namespace Files.App.Filesystem
             }
         }
 
+        public LibraryManager()
+        {
+            InitializeWatcher();
+        }
+
+        private void InitializeWatcher()
+        {
+            if (librariesWatcher is not null)
+                return;
+
+            librariesWatcher = new FileSystemWatcher
+            {
+                Path = ShellLibraryItem.LibrariesPath,
+                Filter = "*" + ShellLibraryItem.EXTENSION,
+                NotifyFilter = NotifyFilters.Attributes | NotifyFilters.LastWrite | NotifyFilters.FileName,
+                IncludeSubdirectories = false,
+            };
+
+            librariesWatcher.Created += OnLibraryChanged;
+            librariesWatcher.Changed += OnLibraryChanged;
+            librariesWatcher.Deleted += OnLibraryChanged;
+            librariesWatcher.Renamed += OnLibraryRenamed;
+
+            librariesWatcher.EnableRaisingEvents = true;
+        }
+
+        public static bool IsDefaultLibrary(string libraryFilePath)
+        {
+            // TODO: try to find a better way for this
+            switch (Path.GetFileNameWithoutExtension(libraryFilePath))
+            {
+                case "CameraRoll":
+                case "Documents":
+                case "Music":
+                case "Pictures":
+                case "SavedPictures":
+                case "Videos":
+                    return true;
+
+                default:
+                    return false;
+            }
+        }
+
+        /// <summary>
+        /// Get libraries of the current user with the help of the FullTrust process.
+        /// </summary>
+        /// <returns>List of library items</returns>
+        public static async Task<List<LibraryLocationItem>> ListUserLibraries()
+        {
+            var libraries = await Win32API.StartSTATask(() =>
+            {
+                try
+                {
+                    var libraryItems = new List<ShellLibraryItem>();
+                    // https://docs.microsoft.com/en-us/windows/win32/search/-search-win7-development-scenarios#library-descriptions
+                    var libFiles = Directory.EnumerateFiles(ShellLibraryItem.LibrariesPath, "*" + ShellLibraryItem.EXTENSION);
+                    foreach (var libFile in libFiles)
+                    {
+                        using var shellItem = new ShellLibrary2(Shell32.ShellUtil.GetShellItemForPath(libFile), true);
+                        if (shellItem is ShellLibrary2 library)
+                        {
+                            libraryItems.Add(ShellFolderExtensions.GetShellLibraryItem(library, libFile));
+                        }
+                    }
+                    return libraryItems;
+                }
+                catch (Exception e)
+                {
+                    App.Logger.Warn(e);
+                }
+
+                return new();
+            });
+
+            return libraries.Select(lib => new LibraryLocationItem(lib)).ToList();
+        }
+
         public async Task UpdateLibrariesAsync()
         {
             lock (libraries)
             {
                 libraries.Clear();
             }
-            var libs = await LibraryHelper.ListUserLibraries();
+            var libs = await ListUserLibraries();
             if (libs is not null)
             {
                 libs.Sort();
@@ -42,35 +139,6 @@ namespace Files.App.Filesystem
                 }
             }
             DataChanged?.Invoke(SectionType.Library, new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Reset));
-        }
-
-        public Task HandleWin32LibraryEvent(ShellLibraryItem library, string oldPath)
-        {
-            string path = oldPath;
-            if (string.IsNullOrEmpty(oldPath))
-            {
-                path = library?.FullPath;
-            }
-            var changedLibrary = Libraries.FirstOrDefault(l => string.Equals(l.Path, path, StringComparison.OrdinalIgnoreCase));
-            if (changedLibrary is not null)
-            {
-                lock (libraries)
-                {
-                    libraries.Remove(changedLibrary);
-                }
-                DataChanged?.Invoke(SectionType.Library, new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Remove, changedLibrary));
-            }
-            // library is null in case it was deleted
-            if (library is not null && !Libraries.Any(x => x.Path == library.FullPath))
-            {
-                var libItem = new LibraryLocationItem(library);
-                lock (libraries)
-                {
-                    libraries.Add(libItem);
-                }
-                DataChanged?.Invoke(SectionType.Library, new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Add, libItem));
-            }
-            return Task.CompletedTask;
         }
 
         public bool TryGetLibrary(string path, out LibraryLocationItem library)
@@ -84,13 +152,34 @@ namespace Files.App.Filesystem
             return library is not null;
         }
 
+        /// <summary>
+        /// Create a new library with the specified name.
+        /// </summary>
+        /// <param name="name">The name of the new library (must be unique)</param>
+        /// <returns>The new library if successfully created</returns>
         public async Task<bool> CreateNewLibrary(string name)
         {
-            if (!CanCreateLibrary(name).result)
-            {
+            if (string.IsNullOrWhiteSpace(name) || !CanCreateLibrary(name).result)
                 return false;
-            }
-            var newLib = await LibraryHelper.CreateLibrary(name);
+
+            var newLib = new LibraryLocationItem(await Win32API.StartSTATask(() =>
+            {
+                try
+                {
+                    using var library = new ShellLibrary2(name, Shell32.KNOWNFOLDERID.FOLDERID_Libraries, false);
+                    library.Folders.Add(ShellItem.Open(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments))); // Add default folder so it's not empty
+                    library.Commit();
+                    library.Reload();
+                    return Task.FromResult(ShellFolderExtensions.GetShellLibraryItem(library, library.GetDisplayName(ShellItemDisplayString.DesktopAbsoluteParsing)));
+                }
+                catch (Exception e)
+                {
+                    App.Logger.Warn(e);
+                }
+
+                return Task.FromResult<ShellLibraryItem>(null);
+            }));
+
             if (newLib is not null)
             {
                 lock (libraries)
@@ -103,9 +192,76 @@ namespace Files.App.Filesystem
             return false;
         }
 
+        /// <summary>
+        /// Update library details.
+        /// </summary>
+        /// <param name="libraryFilePath">Library file path</param>
+        /// <param name="defaultSaveFolder">Update the default save folder or null to keep current</param>
+        /// <param name="folders">Update the library folders or null to keep current</param>
+        /// <param name="isPinned">Update the library pinned status or null to keep current</param>
+        /// <returns>The new library if successfully updated</returns>
         public async Task<LibraryLocationItem> UpdateLibrary(string libraryPath, string defaultSaveFolder = null, string[] folders = null, bool? isPinned = null)
         {
-            var newLib = await LibraryHelper.UpdateLibrary(libraryPath, defaultSaveFolder, folders, isPinned);
+            if (string.IsNullOrWhiteSpace(libraryPath) || (defaultSaveFolder == null && folders == null && isPinned == null))
+                // Nothing to update
+                return null;
+
+            var item = await Win32API.StartSTATask(() =>
+            {
+                try
+                {
+                    bool updated = false;
+                    using var library = new ShellLibrary2(Shell32.ShellUtil.GetShellItemForPath(libraryPath), false);
+                    if (folders != null)
+                    {
+                        if (folders.Length > 0)
+                        {
+                            var foldersToRemove = library.Folders.Where(f => !folders.Any(folderPath => string.Equals(folderPath, f.FileSystemPath, StringComparison.OrdinalIgnoreCase)));
+                            foreach (var toRemove in foldersToRemove)
+                            {
+                                library.Folders.Remove(toRemove);
+                                updated = true;
+                            }
+                            var foldersToAdd = folders.Distinct(StringComparer.OrdinalIgnoreCase)
+                                                      .Where(folderPath => !library.Folders.Any(f => string.Equals(folderPath, f.FileSystemPath, StringComparison.OrdinalIgnoreCase)))
+                                                      .Select(ShellItem.Open);
+                            foreach (var toAdd in foldersToAdd)
+                            {
+                                library.Folders.Add(toAdd);
+                                updated = true;
+                            }
+                            foreach (var toAdd in foldersToAdd)
+                            {
+                                toAdd.Dispose();
+                            }
+                        }
+                    }
+                    if (defaultSaveFolder != null)
+                    {
+                        library.DefaultSaveFolder = ShellItem.Open(defaultSaveFolder);
+                        updated = true;
+                    }
+                    if (isPinned != null)
+                    {
+                        library.PinnedToNavigationPane = isPinned == true;
+                        updated = true;
+                    }
+                    if (updated)
+                    {
+                        library.Commit();
+                        library.Reload(); // Reload folders list
+                        ShellFolderExtensions.GetShellLibraryItem(library, libraryPath);
+                    }
+                }
+                catch (Exception e)
+                {
+                    App.Logger.Warn(e);
+                }
+
+                return Task.FromResult<ShellLibraryItem>(null);
+            });
+
+            var newLib = item != null ? new LibraryLocationItem(item) : null;
             if (newLib is not null)
             {
                 var libItem = Libraries.FirstOrDefault(l => string.Equals(l.Path, libraryPath, StringComparison.OrdinalIgnoreCase));
@@ -144,6 +300,168 @@ namespace Files.App.Filesystem
             return (true, string.Empty);
         }
 
-        public void Dispose() {}
+        public static async void ShowRestoreDefaultLibrariesDialog()
+        {
+            var dialog = new DynamicDialog(new DynamicDialogViewModel
+            {
+                TitleText = "DialogRestoreLibrariesTitleText".GetLocalizedResource(),
+                SubtitleText = "DialogRestoreLibrariesSubtitleText".GetLocalizedResource(),
+                PrimaryButtonText = "DialogRestoreLibrariesButtonText".GetLocalizedResource(),
+                CloseButtonText = "Cancel".GetLocalizedResource(),
+                PrimaryButtonAction = async (vm, e) =>
+                {
+                    await ContextMenu.InvokeVerb("restorelibraries", ShellLibraryItem.LibrariesPath);
+                    await App.LibraryManager.UpdateLibrariesAsync();
+                },
+                CloseButtonAction = (vm, e) => vm.HideDialog(),
+                KeyDownAction = (vm, e) =>
+                {
+                    if (e.Key == VirtualKey.Escape)
+                    {
+                        vm.HideDialog();
+                    }
+                },
+                DynamicButtons = DynamicDialogButtons.Primary | DynamicDialogButtons.Cancel
+            });
+            await dialog.ShowAsync();
+        }
+
+        public static async void ShowCreateNewLibraryDialog()
+        {
+            var inputText = new TextBox
+            {
+                PlaceholderText = "FolderWidgetCreateNewLibraryInputPlaceholderText".GetLocalizedResource()
+            };
+            var tipText = new TextBlock
+            {
+                Text = string.Empty,
+                Visibility = Visibility.Collapsed
+            };
+
+            var dialog = new DynamicDialog(new DynamicDialogViewModel
+            {
+                DisplayControl = new Grid
+                {
+                    Children =
+                    {
+                        new StackPanel
+                        {
+                            Spacing = 4d,
+                            Children =
+                            {
+                                inputText,
+                                tipText
+                            }
+                        }
+                    }
+                },
+                TitleText = "FolderWidgetCreateNewLibraryDialogTitleText".GetLocalizedResource(),
+                SubtitleText = "SideBarCreateNewLibrary/Text".GetLocalizedResource(),
+                PrimaryButtonText = "DialogCreateLibraryButtonText".GetLocalizedResource(),
+                CloseButtonText = "Cancel".GetLocalizedResource(),
+                PrimaryButtonAction = async (vm, e) =>
+                {
+                    var (result, reason) = App.LibraryManager.CanCreateLibrary(inputText.Text);
+                    tipText.Text = reason;
+                    tipText.Visibility = result ? Visibility.Collapsed : Visibility.Visible;
+                    if (!result)
+                    {
+                        e.Cancel = true;
+                        return;
+                    }
+                    await App.LibraryManager.CreateNewLibrary(inputText.Text);
+                },
+                CloseButtonAction = (vm, e) =>
+                {
+                    vm.HideDialog();
+                },
+                KeyDownAction = async (vm, e) =>
+                {
+                    if (e.Key == VirtualKey.Enter)
+                    {
+                        await App.LibraryManager.CreateNewLibrary(inputText.Text);
+                    }
+                    else if (e.Key == VirtualKey.Escape)
+                    {
+                        vm.HideDialog();
+                    }
+                },
+                DynamicButtons = DynamicDialogButtons.Primary | DynamicDialogButtons.Cancel
+            });
+            await dialog.ShowAsync();
+        }
+
+        private async void OnLibraryChanged(WatcherChangeTypes changeType, string oldPath, string newPath)
+        {
+            if (newPath != null && (!newPath.ToLowerInvariant().EndsWith(ShellLibraryItem.EXTENSION, StringComparison.Ordinal) || !File.Exists(newPath)))
+            {
+                System.Diagnostics.Debug.WriteLine($"Ignored library event: {changeType}, {oldPath} -> {newPath}");
+                return;
+            }
+
+            System.Diagnostics.Debug.WriteLine($"Library event: {changeType}, {oldPath} -> {newPath}");
+
+            if (!changeType.HasFlag(WatcherChangeTypes.Deleted))
+            {
+                var library = SafetyExtensions.IgnoreExceptions(() => new ShellLibrary2(Shell32.ShellUtil.GetShellItemForPath(newPath), true));
+                if (library == null)
+                {
+                    App.Logger.Warn($"Failed to open library after {changeType}: {newPath}");
+                    return;
+                }
+
+                var library1 = ShellFolderExtensions.GetShellLibraryItem(library, newPath);
+
+                string? path = oldPath;
+                if (string.IsNullOrEmpty(oldPath))
+                {
+                    path = library1?.FullPath;
+                }
+                var changedLibrary = Libraries.FirstOrDefault(l => string.Equals(l.Path, path, StringComparison.OrdinalIgnoreCase));
+                if (changedLibrary is not null)
+                {
+                    lock (libraries)
+                    {
+                        libraries.Remove(changedLibrary);
+                    }
+                    DataChanged?.Invoke(SectionType.Library, new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Remove, changedLibrary));
+                }
+                // library is null in case it was deleted
+                if (library is not null && !Libraries.Any(x => x.Path == library1?.FullPath))
+                {
+                    var libItem = new LibraryLocationItem(library1);
+                    lock (libraries)
+                    {
+                        libraries.Add(libItem);
+                    }
+                    DataChanged?.Invoke(SectionType.Library, new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Add, libItem));
+                }
+
+                library?.Dispose();
+            }
+        }
+
+        private void OnLibraryChanged(object sender, FileSystemEventArgs e)
+        {
+            switch (e.ChangeType)
+            {
+                case WatcherChangeTypes.Created:
+                case WatcherChangeTypes.Changed:
+                    OnLibraryChanged(e.ChangeType, e.FullPath, e.FullPath);
+                    break;
+                case WatcherChangeTypes.Deleted:
+                    OnLibraryChanged(e.ChangeType, e.FullPath, null);
+                    break;
+            }
+        }
+
+        private void OnLibraryRenamed(object sender, RenamedEventArgs e)
+            => OnLibraryChanged(e.ChangeType, e.OldFullPath, e.FullPath);
+
+        public static bool IsLibraryPath(string path) 
+            => !string.IsNullOrEmpty(path) && path.EndsWith(ShellLibraryItem.EXTENSION, StringComparison.OrdinalIgnoreCase);
+
+        public void Dispose()
+            => librariesWatcher?.Dispose();
     }
 }
