+++ src/Files.FullTrust/MessageHandlers/LibrariesHandler.cs
@@ -96,123 +96,7 @@ namespace Files.FullTrust.MessageHandlers
 
         private async Task HandleShellLibraryMessage(Dictionary<string, JsonElement> message)
         {
-            switch (message["action"].GetString())
-            {
-                case "Enumerate":
-                    // Read library information and send response to UWP
-                    var enumerateResponse = await Win32API.StartSTATask(() =>
-                    {
-                        var response = new ValueSet();
-                        try
-                        {
-                            var libraryItems = new List<ShellLibraryItem>();
-                            // https://docs.microsoft.com/en-us/windows/win32/search/-search-win7-development-scenarios#library-descriptions
-                            var libFiles = Directory.EnumerateFiles(ShellLibraryItem.LibrariesPath, "*" + ShellLibraryItem.EXTENSION);
-                            foreach (var libFile in libFiles)
-                            {
-                                using var shellItem = new ShellLibrary2(Shell32.ShellUtil.GetShellItemForPath(libFile), true);
-                                if (shellItem is ShellLibrary2 library)
-                                {
-                                    libraryItems.Add(ShellFolderExtensions.GetShellLibraryItem(library, libFile));
-                                }
-                            }
-                            response.Add("Enumerate", JsonSerializer.Serialize(libraryItems));
-                        }
-                        catch (Exception e)
-                        {
-                            Program.Logger.Warn(e);
-                        }
-                        return response;
-                    });
-                    await Win32API.SendMessageAsync(connection, enumerateResponse, message.Get("RequestID", defaultJson).GetString());
-                    break;
-
-                case "Create":
-                    // Try create new library with the specified name and send response to UWP
-                    var createResponse = await Win32API.StartSTATask(() =>
-                    {
-                        var response = new ValueSet();
-                        try
-                        {
-                            using var library = new ShellLibrary2(message["library"].GetString(), Shell32.KNOWNFOLDERID.FOLDERID_Libraries, false);
-                            library.Folders.Add(ShellItem.Open(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments))); // Add default folder so it's not empty
-                            library.Commit();
-                            library.Reload();
-                            response.Add("Create", JsonSerializer.Serialize(ShellFolderExtensions.GetShellLibraryItem(library, library.GetDisplayName(ShellItemDisplayString.DesktopAbsoluteParsing))));
-                        }
-                        catch (Exception e)
-                        {
-                            Program.Logger.Warn(e);
-                        }
-                        return response;
-                    });
-                    await Win32API.SendMessageAsync(connection, createResponse, message.Get("RequestID", defaultJson).GetString());
-                    break;
-
-                case "Update":
-                    // Update details of the specified library and send response to UWP
-                    var updateResponse = await Win32API.StartSTATask(() =>
-                    {
-                        var response = new ValueSet();
-                        try
-                        {
-                            var folders = message.ContainsKey("folders") ? JsonSerializer.Deserialize<string[]>(message["folders"].GetString()) : null;
-                            var defaultSaveFolder = message.Get("defaultSaveFolder", defaultJson).GetString();
-                            var isPinned = message.Get("isPinned", defaultJson).GetBoolean();
-
-                            bool updated = false;
-                            var libPath = message["library"].GetString();
-                            using var library = new ShellLibrary2(Shell32.ShellUtil.GetShellItemForPath(libPath), false);
-                            if (folders != null)
-                            {
-                                if (folders.Length > 0)
-                                {
-                                    var foldersToRemove = library.Folders.Where(f => !folders.Any(folderPath => string.Equals(folderPath, f.FileSystemPath, StringComparison.OrdinalIgnoreCase)));
-                                    foreach (var toRemove in foldersToRemove)
-                                    {
-                                        library.Folders.Remove(toRemove);
-                                        updated = true;
-                                    }
-                                    var foldersToAdd = folders.Distinct(StringComparer.OrdinalIgnoreCase)
-                                                              .Where(folderPath => !library.Folders.Any(f => string.Equals(folderPath, f.FileSystemPath, StringComparison.OrdinalIgnoreCase)))
-                                                              .Select(ShellItem.Open);
-                                    foreach (var toAdd in foldersToAdd)
-                                    {
-                                        library.Folders.Add(toAdd);
-                                        updated = true;
-                                    }
-                                    foreach (var toAdd in foldersToAdd)
-                                    {
-                                        toAdd.Dispose();
-                                    }
-                                }
-                            }
-                            if (defaultSaveFolder != null)
-                            {
-                                library.DefaultSaveFolder = ShellItem.Open(defaultSaveFolder);
-                                updated = true;
-                            }
-                            if (isPinned != null)
-                            {
-                                library.PinnedToNavigationPane = isPinned == true;
-                                updated = true;
-                            }
-                            if (updated)
-                            {
-                                library.Commit();
-                                library.Reload(); // Reload folders list
-                                response.Add("Update", JsonSerializer.Serialize(ShellFolderExtensions.GetShellLibraryItem(library, libPath)));
-                            }
-                        }
-                        catch (Exception e)
-                        {
-                            Program.Logger.Warn(e);
-                        }
-                        return response;
-                    });
-                    await Win32API.SendMessageAsync(connection, updateResponse, message.Get("RequestID", defaultJson).GetString());
-                    break;
-            }
+            
         }
 
         protected override void Dispose(bool disposing)
